refactor all providers spi
