+++ README.md
@@ -1,1 +1,10 @@
+<!---
+This file is auto-generate by a github hook please modify r.md if you don't want to loose your work
+-->
+![Build Status](https://github.com/SOFTNETWORK-APP/generic-notification-api/workflows/Build/badge.svg)
+[![Scala Steward badge](https://img.shields.io/badge/Scala_Steward-helping-blue.svg?style=flat&logo=data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAA4AAAAQCAMAAAARSr4IAAAAVFBMVEUAAACHjojlOy5NWlrKzcYRKjGFjIbp293YycuLa3pYY2LSqql4f3pCUFTgSjNodYRmcXUsPD/NTTbjRS+2jomhgnzNc223cGvZS0HaSD0XLjbaSjElhIr+AAAAAXRSTlMAQObYZgAAAHlJREFUCNdNyosOwyAIhWHAQS1Vt7a77/3fcxxdmv0xwmckutAR1nkm4ggbyEcg/wWmlGLDAA3oL50xi6fk5ffZ3E2E3QfZDCcCN2YtbEWZt+Drc6u6rlqv7Uk0LdKqqr5rk2UCRXOk0vmQKGfc94nOJyQjouF9H/wCc9gECEYfONoAAAAASUVORK5CYII=)](https://scala-steward.org)
+[![codecov](https://codecov.io/gh/SOFTNETWORK-APP/generic-notification-api/branch/main/graph/badge.svg)](https://codecov.io/gh/SOFTNETWORK-APP/generic-notification-api/)
+[![Codacy Badge](https://app.codacy.com/project/badge/Grade/f79729e67cce45aba81e1950b91ef8eb)](https://www.codacy.com/gh/SOFTNETWORK-APP/generic-notification-api/dashboard?utm_source=github.com&amp;utm_medium=referral&amp;utm_content=SOFTNETWORK-APP/generic-notification-api&amp;utm_campaign=Badge_Grade)
+[![License: MIT](https://img.shields.io/badge/License-MIT-yellow.svg)](https://opensource.org/licenses/MIT)
+
 # generic-notification-api
