+++ core/src/main/scala/app/softnetwork/notification/handlers/NotificationHandler.scala
@@ -4,7 +4,7 @@ import akka.cluster.sharding.typed.scaladsl.EntityTypeKey
 import app.softnetwork.persistence.typed.scaladsl.EntityPattern
 import app.softnetwork.persistence.typed.CommandTypeKey
 import app.softnetwork.notification.message._
-import app.softnetwork.notification.peristence.typed._
+import app.softnetwork.notification.persistence.typed._
 
 import scala.reflect.ClassTag
 
@@ -18,15 +18,6 @@ trait AllNotificationsTypeKey extends CommandTypeKey[NotificationCommand] {
     AllNotificationsBehavior.TypeKey
 }
 
-trait MockAllNotificationsTypeKey extends CommandTypeKey[NotificationCommand] {
-  override def TypeKey(implicit
-    tTag: ClassTag[NotificationCommand]
-  ): EntityTypeKey[NotificationCommand] =
-    MockAllNotificationsBehavior.TypeKey
-}
-
 trait NotificationHandler
     extends EntityPattern[NotificationCommand, NotificationCommandResult]
     with AllNotificationsTypeKey
-
-trait MockNotificationHandler extends NotificationHandler with MockAllNotificationsTypeKey
