+++ core/src/main/scala/app/softnetwork/notification/persistence/query/NotificationCommandProcessorStream.scala
@@ -1,4 +1,4 @@
-package app.softnetwork.notification.peristence.query
+package app.softnetwork.notification.persistence.query
 
 import akka.Done
 import akka.actor.typed.eventstream.EventStream.Publish
