+++ core/src/main/scala/app/softnetwork/notification/spi/DefaultMailAndSMSAndFcmAndIosProvider.scala
@@ -1,0 +1,7 @@
+package app.softnetwork.notification.spi
+
+trait DefaultMailAndSMSAndFcmAndIosProvider
+    extends MailAndSMSAndFcmAndIosProvider
+    with SimpleMailProvider
+    with SMSModeProvider
+    with FcmAndApnsProvider
