+++ core/src/main/scala/app/softnetwork/notification/spi/FcmAndApnsProvider.scala
@@ -1,0 +1,3 @@
+package app.softnetwork.notification.spi
+
+trait FcmAndApnsProvider extends AndroidAndIosProvider with FcmProvider with ApnsProvider
