+++ testkit/src/main/scala/app/softnetwork/notification/persistence/typed/MockAllNotificationsBehavior.scala
@@ -1,0 +1,9 @@
+package app.softnetwork.notification.persistence.typed
+
+import app.softnetwork.notification.spi.MockNotificationProvider
+
+trait MockAllNotificationsBehavior extends AllNotificationsBehavior with MockNotificationProvider {
+  override val persistenceId = "MockNotification"
+}
+
+object MockAllNotificationsBehavior extends MockAllNotificationsBehavior
